feat(top nav): refactor top nav components , add user menu component
