+++ src/components/TopAppBar/components/SidebarMenuIcon/SidebarMenuIcon.style.ts
@@ -1,4 +1,4 @@
-import { Theme } from '../../../theme';
+import { Theme } from '../../../../theme';
 import { css, SerializedStyles } from '@emotion/core';
 
 const iconWrapper = (theme: Theme): SerializedStyles => css`
