+++ src/components/TopAppBar/components/SidebarMenuIcon/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SidebarMenuIcon';
