+++ src/components/TopAppBar/components/UserMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserMenu';
