+++ src/components/TopAppBar/components/index.ts
@@ -1,2 +1,3 @@
-export { default as LogoPlaceholder } from './Logo.wrapper';
+export { default as LogoPlaceholder } from './Logo';
 export { default as SidebarMenuIcon } from './SidebarMenuIcon';
+export { default as UserMenu } from './UserMenu';
